Try installing Qt via apt beforehand
I have a feeling this will work because it'll pull in mesa and such. It'd be better to not expect Qt to be installed on user's machines.

I really need to check out the qutebrowser travis setup. They get backtraces when tests cause segfaults, which would probably be really helpful in these situations.